Acrescentado testes

+++ src/main/java/Projeto/TestesAutomatizados/service/service/ExcluirSerieService.java
@@ -1,10 +1,10 @@
 package Projeto.TestesAutomatizados.service.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
+import Projeto.TestesAutomatizados.service.exception.SerieNaoEncontradaException;
 import Projeto.TestesAutomatizados.service.repository.SeriesRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
 
 @Service
 public class ExcluirSerieService {
@@ -12,18 +12,21 @@ public class ExcluirSerieService {
     private static final Logger logger = LoggerFactory.getLogger(ExcluirSerieService.class);
     private final SeriesRepository repository;
 
-    @Autowired
-    public ExcluirSerieService(SeriesRepository repository, BuscarSeriesService buscarSeriesService) {
+    public ExcluirSerieService(SeriesRepository repository) {
         this.repository = repository;
     }
 
     public void excluir(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("ID não pode ser nulo");
+        }
+
         repository.findById(id).ifPresentOrElse(serie -> {
             repository.delete(serie);
             logger.info("Série com ID {} excluída com sucesso.", id);
         }, () -> {
             logger.error("Erro: Série com ID {} não encontrada.", id);
-            throw new RuntimeException("Série com ID " + id + " não encontrada.");
+            throw new SerieNaoEncontradaException("Série com ID " + id + " não encontrada.");
         });
     }
 }
